+++ src/components/EditSections/EditReadingRoomAccess/EditReadingRoomAccess.js
@@ -1,7 +1,8 @@
-import React, { useEffect } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { noop } from 'lodash';
+import { cloneDeep, noop, orderBy } from 'lodash';
+import { OnChange } from 'react-final-form-listeners';
 
 import {
   Accordion,
@@ -10,8 +11,12 @@ import {
   MultiColumnList,
 } from '@folio/stripes/components';
 
-import { rraColumns } from './constants';
+import { rraColumns, DEFAULT_SORT_ORDER } from './constants';
 import { getFormatter } from './getFormatter';
+import { sortTypes } from '../../../constants';
+import { getReadingRoomSortedData } from '../../util/util';
+
+const rraList = 'readingRoomsAccessList';
 
 const EditReadingRoomAccess = ({
   expanded,
@@ -30,30 +35,73 @@ const EditReadingRoomAccess = ({
     [rraColumns.ACCESS]: '15%',
     [rraColumns.READING_ROOM_NAME]: '25%',
   };
+  const sortInitialState = {
+    data: [],
+    order: DEFAULT_SORT_ORDER,
+    direction: sortTypes.ASC,
+  };
+  const [sortedRecordsDetails, setSortedRecordsDetails] = useState(sortInitialState);
 
   useEffect(() => {
-    const unregisterReadingRoomAccessList = form.registerField('readingRoomsAccessList', noop, { initialValue: [] });
+    const unregisterReadingRoomAccessList = form.registerField(rraList, noop, { initialValue: [] });
     return () => {
       unregisterReadingRoomAccessList();
     };
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
+  useEffect(() => {
+    setSortedRecordsDetails(prev => ({
+      ...prev,
+      data: orderBy(formData, [data => data[prev.order]?.toLowerCase()], prev.direction)
+    }));
+  }, [formData]);
+
+  const updateSortedRecordDetails = useCallback((updatedRRAFormValues) => {
+    if (updatedRRAFormValues?.length) {
+      const recordDetails = cloneDeep(sortedRecordsDetails);
+      updatedRRAFormValues?.forEach(updatedRecord => {
+        const index = recordDetails?.data?.findIndex(record => record.readingRoomId === updatedRecord.readingRoomId);
+        if (index !== -1) {
+          recordDetails.data[index] = updatedRecord;
+        }
+      });
+      setSortedRecordsDetails(recordDetails);
+    }
+  }, [sortedRecordsDetails]);
+
+  const { data: sortedData, order: sortOrder, direction: sortDirection } = sortedRecordsDetails;
+
   return (
     <Accordion
       open={expanded}
       id={accordionId}
       onToggle={onToggle}
-      label={<Headline size="large" tag="h3"><FormattedMessage id="ui-users.readingRoom.readingRoomAccess" /></Headline>}
+      label={
+        <Headline size="large" tag="h3">
+          <FormattedMessage id="ui-users.readingRoom.readingRoomAccess" />
+        </Headline>
+      }
       displayWhenClosed={<Badge>{formData.length}</Badge>}
     >
+      <OnChange name={rraList}>
+        {(updatedRRAFormValues) => {
+          updateSortedRecordDetails(updatedRRAFormValues);
+        }}
+      </OnChange>
       <MultiColumnList
         striped
-        contentData={formData}
+        contentData={sortedData}
         columnMapping={columnMapping}
         visibleColumns={visibleColumns}
         formatter={getFormatter(form)}
         columnWidths={columnWidths}
+        sortOrder={sortOrder}
+        sortDirection={`${sortDirection}ending`}
+        onHeaderClick={(e, meta) => setSortedRecordsDetails(
+          getReadingRoomSortedData(e, meta, sortedRecordsDetails)
+        )}
+        sortedColumn={sortOrder}
       />
     </Accordion>
   );