+++ CHANGELOG.md
@@ -23,6 +23,9 @@
 * Improve UX of Reading Room Access accordion selection box in user profile edit. Refs UIU-3151.
 * Set Users date format in all date input fields as per Session Locale Settings. Refs UIU-2879.
 * Make unblocked fee-fine actions if item is claimed returned and loan is closed. Refs UIU-3125.
+* Add the possibility to use "pay" and "waive" fee/fine permissions separately. Refs UIU-3109.
+* Use showing/hiding approach for fee/fine action buttons. Refs UIU-3156.
+* Support sorting for each column in the Reading Room Access accordion in User Profile Edit. Refs UIU-3132.
 * Add the 'Preferred email communication' field in the Contact information accordion of the user profile. Refs UIU-3152.
 
 ## [10.1.1](https://github.com/folio-org/ui-users/tree/v10.1.1) (2024-05-07)